Include an initial Readme file
This is an initial set of documentation that tries to describe what
the plugin is supposed to do.

Probably in the future this could be edited considerably, but I'm also
using it at the moment as a way to keep thoughts straight on exactly
what and how we want this thing to work.