Initial package info classes
This includes a class PackageInfo that is meant to wrap all of the
dirty details of known packages, and a PackageList class that holds
a dictionary of PackageInfo instances.

The PackageInfo class holds the name of the package as well as the
location where the shipped, installed or unpacked version of the
package is contained. Any one of the three may be set.

The PackageList class scans for all possible plugins, creating or
filling out PackageInfo classes as needed, and provides some internal
methods to allow for getting at the found data easier.