Document/Enhance packages.py
This adds class documentation comments to the classes in
lib/packages.py that explains their purpose.

Additionally, when a sublime-package is found, we obtain the last
modification time for that package right away since that is likely to
be useful. We don't do this for the contents of an unpacked package
because that could be a lengthener operation.

Possibly we want to redact this ability and do it on demand; I don't
know if getting the modification time of a file is a time concern at
all, even for people with an excessive list of packages.