Include deletion of existing text
This selects everything in an existing buffer and deletes it prior to
generating the new input. We also scroll back to the top of the buffer
after the insert so that we can view the table from the start.

Additionally I included the closing header line in the table that I
forgot last time, so the table looks correct again.