Include self contained method for output to a view
This uses an enhanced version of the code that is currently in the
override_audit command file, making it more general purpose.

The new method can send output to a view, either reusing the one with
the same title as is being used, or creating a new one. A syntax can
be applied as well as arbitrary settings, and the buffer can be
cleared if desired.

In the case that the buffer is not cleared, the code takes care to
retain the current selection information and viewport position and
put it back to that state later.