Rework the layout
The table is less table-like now. It's still got separators and such,
but now instead of named columes the columns are blind and the
characters in the "[ ]" marks represent [S]hipped, [I]installed or
[U]npacked.

There is also now a summary of the counts of each, including disabled
packages, whose names are output in square brackets so that we know
what they are.