Add dependency detection
As far as I can tell from looking at documentation and implementation,
PackageControl always installs a dependency as an unpacked file, and
further always puts a specific metadata file inside.

Using that info, the code now detects when a package is a dependency
by seeing if there is a file by that name when an unpacked path is
added.

The PackageList tracks the number of items that are dependencies as
well.