Enhance syntax and layout
This change swaps the table layout to have the package name first and
properties second, and provides a new syntax for everything.

The syntax takes care to assign "entity.package.name" to all package
names, along with (currently) meta.toc-list to ensure that they end
up in the symbol list.

Additionally, [Packages] are disabled and <Package> are dependencies,
and each is still captured by name (excluding the symbol) with an
additional suffix of .disabled or .dependency, while normal packages
have a suffix of .enabled (to make them distinct).

With the new syntax in place, disabled packages appear as a comment
and dependencies appear as storage modifiers, while regular packages
are named entities.

The inclusion of the scopes allows key bindings to be made that can
trigger inside of a specific package type.