Fix Installed Packages scanning
I have no idea what the use case might be, but Subline recursively
looks for sublime-package files in the InstalledPackages folder, not
just inside of InstalledPackages itself, so we should probably do the
same when we're gathering the packag list.

I'm not sure if this is stricly needed or not, because PackageControl
doesn't list packages that are in subfolders of InstalledPackages and
neither does PackageResourceViewer, but nevertheless the Sublime
console shows the package being loaded and unloaded when you add and
remove it.