Add extra PackageInfo accessors
Working towards getting the contents of a package from a packed
package file, we need to be able to determine which file  will be
used, since if there is an installed package with the same name as a
shipped package, Sublime ignored the shipped version.

This also adds a simple check for determining if a package has any
unpacked files, just to make things a little more readable.