Add API for collecting package contents
This new API is capable of collecting the contents of both a packed
sublime-package file as well as the contents of a loose package
directory.

Internally there are routines to do the work and memoize the results
and some simple accessor methods for collecting the input based on any
of the possible paths that the package data can be contained in.

In order to facilitate operation on all supported Sublime operating
systems, the code uses os.normcase to normalize all of the filenames
found, so that (later) when we do comparisons to check for overrides,
we don't get kicked in the jimmies.