Insert method for detecting potential overrides
In the interests of speed, this new method just checks to see if an
override of a certain type (simple or not, see below) is at least
possible.

There are two "types" of overrides. Here a simple override means that
there is an unpacked directory for this package, and so it may contain
a file that overrides something in the package. This is the most
common type of override, which is why this is the default mode.

A "complex" override happens when there is an installed
sublime-package file that has the same name as a shipped package file,
e.g. ~/.config/sublime-text-3/InstalledPackages.XML.sublime-package.

In this case, Sublime will ignore the shipped version of the package
file and use the installed version instead.

This is a less likely form of override, but is still detected at the
user's request, since that could be a source of potential woes.