Insert the ability to detect overriden files
This includes an override_files() method which ties all of today's
changes together.

Given a the same boolean value as the previous methods to determine a
simple or complex override, this will collect the list of files of the
underlying package and the unpacked contents, and then tell you which
of those files is an override.

This does not do the checking required to determine if an override is
out of date or not; it just tells you what files are being overriden,
which may be an empty list of files if it's not possible for such an
override to exist or the unpacked directory does not have any files
that the underlying package contains.