Implement command to show overrides
This is still crude, but basically it is a stripped version of the
other command. Here a package only shows up in the list if it has any
overrides, and both kinds of overrides are displayed; it should be
either/or as a config option.

The package names are decorated like in the package list, with a
shortened version of the Shipped/Installed/Unpacked indicators for
debugging purposes.

Also, there is no syntax.