Tweak text additions to buffers
As originally written, the append method used the "insert" command to
insert the text. This works, but it has a couple of drawbacks:

 * You have to swap the cursor position to the end of the buffer
 * auto indent can kick you in the nads

Lord knows why this didn't occur to me initially, but the append
command always appends at the end and does not seem to be subject to
indenting like the other method is, so its a double win.