Include diff API in PackageInfo
This adds a new public method, override_diff(), which takes the name
of an override file and a number of context lines, and attempts to
calculate a unified diff showing how the override differs from the
underlying file.

This diff is currently always against an override in the Packages
directory; is it not currently possible to diff an override that comes
from an installed sublime-package overriding a shipped package.

As a consequence, the source of the diff always comes from a
sublime- package file; the version in Installed Packages if one is
there, otherwise the one that ships with Sublime in the executable
directory.

The other side of the diff is taken from the unpacked package folder,
as the only place that an override can come from currently.