Implement new diff_context_lines setting
This setting (which currently defaults to 3) controls the number of
context lines the diff commands will use when displaying diff output.

The value of 3 comes from the difflib which uses that value as its
default, but this is configurable so that people can have as much
context as they like (I generally like a little bit more in this
style of diff).