Implement overide_audit_diff_override command
Currently, this command takes as an arguments package and a file that
is part of that package, and attempts to open a diff view of the
override.

This tries to catch all possible errors (no such package, no such file
in the package, no override for that file in the package, etc) and act
as appropriate.

Assuming that the arguments specify an actual override, the command
will perform a unified diff and display the output in a new view,
assigning the diff syntax to it so that Sublime will handle it as it
does its own internal diffs.