Fix packed package file contents being read in the wrong encoding
By default, io.TextIOWrapper uses locale.getpreferredencoding(False) as the encoding, which will be different on every system. As the files are always UTF-8, explicitly tell it to use the correct encoding.
Specifically, without this change, problems can occur reading .sublime-menu files, because they contain ellipsis characters which might not be present in the user's locale's default encoding.