Make override_audit_diff_override actually useful
This command still takes a "package" and "file" argument, but both are
now optional.

When no package is provided, the command will show a list of all
packages that have at least one override, and after selecting a
package a list will be presented for all of the overrides that exist
in that particular package.

When a package is provided, the command directly shows the list of
overrides within that particular package for you to select from.

In either case, selecting an override from the list invokes the diff
for that particular override.

These options allow for key commands, menu entries or command palette
items custom tailored to any particular use case.

This needs a bit of work still. In particular, when there are no
packages that have overrides or no overrides inside a package (which
can only happen if you specify a package directly), the situation is
sent to the console but that's it.

I'm not sure if the best thing to do is an error dialog, pop up the
panel anyway with a single "No stuff" entry, the status line showing a
message, or some combination thereof.