Swap date format in diff output
Previously the diff output was using the result of ctime() to show the
time stamps of the two files, which looks a bit ugly. This is now
modified to use a ISO 8601 timestamp similar to what "diff -u" does,
although there currently is no time zone information.

Probably this is a great candifate for a setting or something, so that
people could tweak it as desired, but this seems much easier to grok
at a glance.