Fix potential badness on diffing
Although all of the code leading to the diff routine is careful to not
attempt to invoke it when it's not possible to (e.g. there is no
installed sublime-package or no unpacked path) it is possible to
invoke the command with such arguments.

This makes the command block access with an error to the console
instead of triggering a stack trace about trying to append None and a
string or some such.