Serious revamp of the README file
Most of the contents of this file are the same, generally speaking,
but everything has been fleshed out and made more "Wall of Text"-like
(if that is indeed possible).

This better outlines what exactly the package is for and the
terminology it uses.

Main changes (other than copy editing) include the addition of a
license and a description of the available commands and what they're
supposed to do. There is also a section on installation using package
control, which is possible if you manually add the repository to your
install.