Add new option for empty diffs
There is a new configuration option, "diff_unchanged" which controls
what happens when you diff an override but the diff comes out empty
due to no changes.

When the option is set to "ignore" no new view is opened at all. If
it is set to "diff", the (legacy) behaviour of opening up a tab that
says there are no differences is used. Lastly, when set to "open" the
override file will be directly opened.

In all cases, the status bar shows a message that indicates that no
changes were found for the diff.

This does not actually validate that the option is only these three
values currently, since I don't know what the cleanest way to do that
is (assuming people generally do such things). Instead, the code
assumes the "diff" outcome if the setting is set to an unrecognized
value.