Slightly modify the output of override list
This applies to the command that shows the complete list of all known
(and unignored) overrides. Basically this is the same as it was except
that the wording matches the README now (simple and complete as the
two override types), and there is handling for what happens if there
are no packages found with overrides.

There is also a new comment that points out that since a complete
override completely ignores a shipped package in favor of a new one,
showing a list of files that are different between the two is not
entirely useful in all cases.

Certainly if the override was, e.g. a newer version of the shipped
version, this is handy to allow you to see what you're changing. On
the other hand if the use case is to replace a package with an
otherwise empty package or something wildly different, this is
unlikely to show anything interesting.

I don't know if the best way to go would be to have a setting, make
the page more verbose to tell you, or assume that anyone that's doing
that knows what is going on. That's currently a pending to-do item.