Honor "ignore_overrides_in" in diff command
This setting is supposed to cause OverrideAudit to ignore packages for
the purposes of displaying or reporting overrides, so it makes sense
that it should be used in the diff command.

Now packages ignored in this fashion will be excluded from the list of
packages that you can select overrides to diff from. This should have
been implemented originally but I forgot all about it until after the
fact.