Fix diffing on case insensitive OS's
The zipinfo structure that falls out of a ZipFile instance that
represents a sublime-package file contains the name of a file in a
case sensitive manner, even on operating systems that are case
insensitive. Additionally, it is possible for a zip file to contain
multiple entries that have the same file name.

If we add to this the fact that the intersection operation done to
determine what unpacked files represent overrides causes files from
the unpacked package to be listed (instead of those from the zip file
itself), we run into problems.

On a case insensitive operating system (i.e. not Linux), the file
EXec.py is equivalent to exec.py, but since no such entry exists
inside of the sublime-package, the diff fails.

A possible fix for this would be to do a case insensitive scan of the
contents of the zip file to find the appropriate entry, but I don't
know good that is as an idea due to the fact that the same file can
appear multiple times, possibly with a different case.

This fix reverses the order of the sets in the intersection, which
causes the results to be culled from the package contents and not from
the unpacked contents. This ends up in an otherwise identical set
because set intersection is commutative, although the result is
different here due to the (lack of) case comparisons being used.

This also has the side effect of making the resulting output in
reports look more "pleasant" because the case is "correct" on the
filenames.

As far as I am able to tell from perusing documentation and source
archives, this ordering happens on purpose based on the order the sets
are provided in and it is not just an implementation detail, although
I could be wrong, in which case this commit message will be egg on my
face in the (probably not too distant) future when this assumption is
proved wrong.