Rename syntax for the package list
I'm not fully happy with the way I've set up the syntax highlighting
for the package list table, since it is essentially hijacking other
scopes to try and get an appropriate color, but for the time being
I've decided to leave that in for later tweaking.

As such, the syntax should be renamed to better track what it's doing
since there will soon be another one for the override list.