Change format of the override listing
The report that shows you the list of all packages is now a littler
clearer and has a syntax assigned to it to colorize it a bit. This is
similar to the previous version except some redundancies have been
removed.

The addition of square or angle brackets to the package name to
represent a package that is a dependency or disabled has been removed
since a dependency is always unpacked and thus does not support the
idea of an override and we skip disabled packages.

Also in the case of a complete override (installed sublime-package
that has the same name as a shipped package) it doesn't seem right to
show the list of files that are being overridden since *every* file is
being overridden in this case.

Instead we just have an indication that this package represents such
an override.