Add overrides to the symbol list
This adds preferences that add the list of overrides for any given
package to the symbol list for the current file, indented. This allows
for quickly filtering and skipping/jumping to a particular override if
desired.

The list of packages was already in the symbol list (as it is in the
package list table) due to the scope applied to it with the syntax.