List Installed Packages becomes Package Report
This renames the command and associated menu/palette entries to match.
Funtionally the command is the same but is having its name changed to
fit better with upcoming changes, and the notion that as an Audit
what we are seeing here are various reports on things (for the most
part).

Also included here is a small format change to the text of the report
to clarify the number counts at the top of the report, making things
seem a little clearer.