Clarify the single override command name
This is just a small tweak that changes the name of the command that
diffsd a single override, which is purely cosmetic.

This also adds a bit of extra documentation to the README to indicate
that there are extra options that control what happens when you
execute the command but there are no differences.