List Package Overrides becomes Override Report
Much like the similar change for List Installed Packages, this changes
the command name and associated menu/command palette text to show that
this command is for generating a report on overrides in packages.

The output of the report itself has been slightly visually tweaked as
it did not gracefully handle the situation in which a package with a
complete override had no simple overrides as well. This is now handled
by ensuring that text that indicates this is generated.