Performance tweak in Override Report
The original version of this command displayed not only the list of
unpacked overrides but also the list of files being overridden in a
complete override; that is to say it would collect the list of files
in the shipped version of the package that were also in the installed
version.

That part of the report was redacted because it's not fully
interesting due to how in this case Sublime completely ignores the
shipped package anyway.

However the code was still collecting that list of overrides even
though it is only neccesary to know if there MIGHT be any, as in this
case the test checks to verify that this package is actually a
complete override, which is what we want anyway.

This change is fairly simple but saves a bit of unneccesary work.