Return output view for later usage
This modifies the output_to_view() function in output_view.py so that
once it has finished generating output, the view that was just used to
do the output is returned to the caller.

This is a rather obvious omission, but in some cases you may want to
do further manipulations on the view after generating the output to
it.