Extra parameters for override_diff
This modifies the override_diff() method of the PackageInfo() class
with a couple of extra helpful (but optional) parameters.

The first of these is empty_result, which if given is set to the
result returned in the case that the diff of the two source files
comes out identical. Previously this return value was just the empty
string.

The second parameter is indent, which if given is a number of spaces
of indent to apply to each line of the resulting diff output. This
includes the result provided by empty_result if the diff is otherwise
empty.