Include bulk diff report
This implements a command which will find all overrides in all
packages that have any and perform a diff on them.

The output of this is sent to a buffer which displays the names of all
packages, followed by the list of all overrides for those packages,
followed by the diff results for that override (if any).

Each level is indented a level deeper than the previous so that it is
possible to fold away the results of entire packages or for individual
overrides (within the confines of Sublime folding).