Generate report when there are no overrides
As originally written, the code would not generate any output to the
output view when told to diff all overrides in all packages if there
were no overrides to diff.

This is a little counter-intuitive UX-wise because there is no
indication that you actually did anything. In order to remain in line
with what happens when you attempt to generate the complete list of
overrides into a buffer, the command now generates output to tell you
that there was nothing to diff.