Fix diff goto symbol in a more intelligent way
The syntax used in the "All Packages" diff is a modified version of
the Packages/Default/Diff.sublime-syntax file, which puts the regions
from the context diff in the symbol list.

I originally removed them by including symbol preferences to redact
them, but as it transpires I'm just blind and didn't notice the  meta
.toc-list that they were scoped with.

This change removes that scope and the (now) extraneous settings file
to make the same change more clear.