Implement bulk diff for a single package
This is either brilliant design or a horrible hack; you get to decide.

In order to allow a bulk diff to occur for only a single package
instead of all packages, this ties into the existing command that
allows for interactively selecting a single override.

The command override_audit_diff_override now takes an optional
argument named "bulk" which defaults to False. When it is true AND no
package is provided, it will open up the list of packages that have
overrides as normal, but when a package is selected instead of showing
you the list of overrides it will instead invoke the recently added
command with the selected package, which causes it to diff only that
package.

In order to fully support this, override_audit_diff_package now titles
the report with the name of the single package it was given instead of
saying how many packages it is diffing.

As an extra case, the command detects whether it is diffing only a
single package because you told it to versus when there is actually
only a single package so that it can accurately title everything with
"All Packages" or the package name provided.