Documentation clarifications
You know how sometimes after you push a release you almost immediately
notice that you forgot to fix some obvious problems, and also you're
not happy with a change and would rather have not done it, but since
you already did the push you have to let it stand and fix it later
instead of ammending the existing commit?

This is like that.