Fix a Windows issue with overrides in subpaths
There was an issue with determining the list of unpacked files that
are overriding a packed file, which stopped such files from appearing
in things such as a bulk diff or even in the panel for choosing an
override to diff manually.

The reason for this is that regardless of the platform that a zip file
is made on, paths are always represented with a unix path separator
character. However, os.path.join() and os.path.relpath() return paths
with a windows path separator on windows.

This causes an issue trying to compare the two lists of files, because
the entries are different due to the disparity in the path characters.

A first blush attempt at fixing this is to use os.path.normpath() on
the path that falls out of the zip file so that it matches what the
system will generate. However this makes the file unfindable when you
try to open it from the zip file.

For expediency this fix takes the tack of simply replacing the path
separators on windows with a unix path separator when gathering local
package files so that they compare (they are always converted to a
normalized case) so that the files can be found.

There is likely some more pythonic way to fix this (my guess would be
to save the zipinfo that falls out when you scan the zip file and use
that to open the file) but I'm going this way at least for the time
being to get everything up and running for an initial release.