Add static PackageInfo.check_potential_override()
This adds a static class method to the PackageInfo class which allows
you to give it a filename and tell you if the file that is represented
could conceivably be an override or not.

The provided filename either needs to be absolute and contain the
unpacked packages path, or be relative (in which case it is assumed to
be relative to the packages path).

This checks to see if there is a sublime-package for the package named
either as a shipped package or anywhere inside of the installed
packages directory.

No check is done that the package actually contains such a file,
hence the "potential" in the name. The caller can verify that on their
own if desired.

The method returns None if there is no chance that this is an
override, or a tuple containing the name of the package and the name
(and subpath, if any) of the override file if the checks pass.