Factor away diffing code
This change pushes the code for executing a diff of a single override
out of the command that performs that task and into a module level
function that does the same task.

In order to support some upcoming changes, the diff method takes a
couple of optional parameters which allow you to force it to always
display the "No changes" text or force it to reuse the existing view
(and clear it) based on the settings.

Also pushed out is a new method for opening the actual override
itself, which is used inside of the diff when the option to open the
file on diff when there are no changes is set.