Include content items for Open/Diff
This uses recent functionality to provide context menu items and a new
command palette entry that allow you to switch between editing  the
contents of an override and a diff of it.

To do this an event handler catches whenever we open or save a file
that appears that it could be an override and attaches known settings
to the view object to tell it what package and override the view
represents.

These are used by a context menu item (within the view and from the
tab bar) to either open a view that allows you to edit the override or
view a diff of it, depending on what state the current view is in.

This allows for quick tweaking of the override, for example.

A similar command is available in the command palette for this
functionality while the current view is an override, with a caption of
"OverrideAudit: Swap Diff/Override View" since items in the command
list cannot set their own captions and I don't want to implement two
commands at this point just to do this.