Fix detection of manually opened override files
The PackageInfo class static method responsible for detecting if a
file is potentially an override was incorrectly asking sublime for
paths at module load time and not after being told that the plugin was
fully loaded.

Initial testing seemed to indicate that this works, but it only works
if sublime is already running at the time that the package is loaded,
so it is entirely unsafe to rely on that.

I presume that I had a bad combination of test code and restarts to
verify that made me think that works when in fact it does not.