Add context menu to open/edit from a report
This extends the context menu items added previously so that they now
operate also via a bulk diff or an override report, basically wherever
OverrideAudit is displaying a list of packages and override names.

This endeavours to use the same command for all three types of context
entries (the one previous and the two new ones) because they're 99%
identical.