Some code cleanup
This includes some concessions to a few of the flake8 conventions.
There is still a little bit of code that violates, mostly to do with
lines that are slightly longer than 80 lines or which have "unpleasing
indent for continuation lines".

As of right now I'm more interested in the code having readable,
understandable variables than on arbitrary restrictions on the length
of a line in this day and age of monitors with sugh ridiculous
resolutions that they physically make my eyes ache.