Include a context menu for bulk diffing a package
This adds in a context menu item which is offered for any package
displayed in an Override Audit report, such as a package list,
override list, diff report, etc.

This is semi-nonsensical in a single bulk diff because the operation
would just regenerate the existing bulk diff for the single package,
but this may come in handy to get a new overview once existing
overrides are edited or removed.