Include support for Ignored package overrides
At some point I must have had a little mild brain infarction because I
decided that ignored packages could not have overrides for some
bizarre reason, and so they were not properly supported.

In particular such packages would show up in the override list and
bulk diff views, but not marked as if they were ignored. The syntaxes
now properly track this, so that visually everything works the way we
want it to.

Dependency packages are still not supported for having overrides, but
in this case this is (I think) correct because as far as I have been
able to determine PackageControl only ever installs packages as
unpacked packages and never as packed ones, presumably so that it can
use standard python mechanisms to load them on behalf of packages,
although that is just a guess.