Fix regression with Diff/Edit context menus
A previous commit broke the context menu entries for being able to
open or diff an override from within an override or diff report due to
the change in scope of package names making it impossible to determine
the appropriate package name.

This is now fixed and the command also detects the situation and
generates an error message instead of throwing an exception.