Code cleanup before some refactoring
Add in the missing Command suffix for some of the more recently added
commands, and reorder the commands to be in a more logical order.

There are (hopefully) some upcoming refactoring changes to seriously
clean up the context specific handling code in preparation for adding
in the deletion code, and I would like the code to be as clean as
possible before starting.