Refactor Override Diff/Edit context command
This re-implements the existing command (with a name and argument
change) to use a new context helper class for determining in what
context the command is being applied.

This uses the newfound knowledge that when used in a tab context menu
the view that the command gets for various commands is not necessarily
the correct view, and in that case the view needs to be looked up by
its group and tab index.

The new context helper class abstracts this away by first determining
what view to apply the command to by first checking if there is a
group and index specified, and looking up is needed.

Next, based on either settings in the view or the buffer content under
the mouse, the package name, override name, and diff status (a
boolean) are extracted.

This allows the new delete command to easily determine in exactly what
context it is being applied and what action it should be taking.

This still uses a single command for either editing, viewing or
toggling the state between a diff or a view, but now the action is
taken by a specific "action" argument which must be one of "toggle",
"edit" or "diff". This makes the code easier to comprehend.

This does mean that for the context menu the same command needs to
be added multiple times, however.