Include override deletion
Include the ability to delete an override via context menu and the
command palette.

All of the locations that offer a context menu now support an entry
which will delete the override from disk, optionally after confirming
that the delete should happen. Confirmation defaults to being on.

This uses the same send2trash that the Default package uses to delete
files.