Fix bulk diff of a single package
Although it is possible to bulk diff a single package, the command
would filter away all packages given that did not have overrides.

This meant that attempts to bulk diff such a package by itself would
cause a report of "0 packages diffed", which while technically
accurate is not entirely useful.

This now modifies that behaviour so that if you attempt to bulk diff a
package, the command follows through anyway (as long as the package
actually exists). This ends you in a report that explicitly tells you
that the package has no overrides, which seems much nicer.