Include Package/Override name in context menu
This simple change modifies the default command description for the
context menu commands on packages and overrides so that the menu text
tells you what package or override the command will affect.

This may come in handy for people using a smaller font as an extra
verification step that the context menu is being invoked in the
correct location when used in an edit buffer, for example.