Fix package reload problem
Based on observation, Sublime will execute plugin_loaded() in any
python file that it initially loads when it initially starts and loads
plugins, but after that point only plugin_loaded() from a top level
module member (i.e. the ones that Sublime automatically loads) will
have its plugin_loaded() method called.

This causes us problems in that packages.py has a plugin_loaded() to
set up the shipped packages static member. This gets invoked at
startup as we expect, but not when the package is reloaded for any
reason, which means that any code that uses that method stops working
after a package reload.

Probably the ultimate solution is some python mysticism that forces
the module to reload every time the core file loads or something along
those lines (if so it should do it for all of them and our development
life gets a whole lot easier).

In the interim I have shifted the code from plugin_loaded() in
packages.py to a static class method in PackageInfo which gets invoked
from the plugin_loaded() of override_audit.py so that it always gets
called no matter what.