Hide package bulk diff context in its own report
All reports which have a package name in them (Package report,
Override report, Bulk diffs of both kinds) include a context menu on
the name of a package that allows for a bulk diff of that package.

This change hides the context menu on a package in a single package
diff report because in that case the option does the same thing as the
"refresh report" context item which is also already there.

This is a small visual tweak only, to make things look "more correct".