Add key binding and context to swap edit/diff view
This adds in a context for use in a key binding to determine when the
current view is a view containing an override (editing it or a diff
of it) so that key bindings can be applied in this context.

Default keys that shadow the keys that Sublime uses by default to swap
between e.g. a header and its source file for certain extensions have
been added.