Include ability to check for expired overrides
This extends the PackageInfo class to examine the list of override
files for any given package (if any) and compare their modification
times to the time of the sublime-package file that Sublime Text is
currently using as the source of the package.

This culls the list of overrides to include only those for which the
source file is newer than the override.

As with the override_files() method, the results of this call are
cached once it is made.

Although the method returns a list of override files that are out of
date for a complete override (an installed sublime-package masking a
shipped one), this is currently implemented by comparing the time
stamps of the two source packages directly and not by comparing the
individual files (which may or may not make sense).

For the case of simple overrides, each unpacked override file has its
timestamp compared to the source package file that Sublime is using,
and not to the last modification timestamp included within the package
itself.